backend/requirements.txt
fastapi[standard]
uvicorn[standard]
anthropic
playwright
supabase
python-dotenv
resend
cryptography
httpx
pydantic>=2.0
pydantic-settings
sse-starlette
apscheduler
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
python-jose[cryptography]
slowapi
email-validator
